Well, this works (sort of)

+++ DBFramework/ActionProvider/SqlDBActionProvider.cs
@@ -10,7 +10,7 @@ namespace DBF
     /// <summary>
     /// 
     /// </summary>
-    public class SqlDBActionProvider : IDBActionProvider
+    public class SqlDBActionProvider : DBActionProvider
     {
         #region Private Members
 
@@ -44,20 +44,20 @@ namespace DBF
         /// <summary>
         /// Gets the <see cref="List{T}"/> from a sql database
         /// </summary>
-        /// <typeparam name="T">The type of data to get from the corresponding table in the <see cref="DBContext"/></typeparam>
+        /// <typeparam name="TModel">The type of data to get from the corresponding table in the <see cref="DBContext"/></typeparam>
         /// <param name="predicate">An <see cref="Action"/> to set the where clause for the select from the sql database</param>
         /// <returns></returns>
-        public async Task<DBSet<T>> Fetch<T>(Action<T> predicate = null) where T : class, new()
+        public async override Task<DBSet<TModel>> Fetch<TModel>(Action<TModel> predicate = null)
         {
             return await Task.Run(() =>
             {
                 // Get the name of the database table
-                string table = (typeof(T).CustomAttributes.Count() > 0) ? (typeof(T).GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : typeof(T).Name;
+                string table = (typeof(TModel).CustomAttributes.Count() > 0) ? (typeof(TModel).GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : typeof(TModel).Name;
 
                 // Get the properties of the given object type
                 string columns = "";
                 List<string> dbcolumns = new List<string>();
-                var props = typeof(T).GetProperties();
+                var props = typeof(TModel).GetProperties();
                 foreach (var prop in props)
                     dbcolumns.Add((prop.CustomAttributes.Count() > 0) ? (prop.GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : prop.Name);
                 var count = 0;
@@ -73,7 +73,7 @@ namespace DBF
                 string where = null;
                 if (predicate != null)
                 {
-                    T item = new T();
+                    TModel item = new TModel();
                     predicate(item);
                     foreach (var prop in item.GetType().GetProperties())
                     {
@@ -108,7 +108,7 @@ namespace DBF
                 // Generate the query
                 string query = $"SELECT {columns} FROM {table} WHERE {where ?? "1 = 1"}";
 
-                DBSet<T> retval = new DBSet<T>();
+                DBSet<TModel> retval = new DBSet<TModel>();
 
                 using (SqlConnection connection = new SqlConnection(_ConnectionString))
                 using (SqlCommand command = new SqlCommand())
@@ -121,13 +121,15 @@ namespace DBF
                     {
                         while (reader.Read())
                         {
-                            T item = new T();
+                            TModel item = new TModel();
                             foreach (var prop in props)
                             {
                                 string propname = (prop.CustomAttributes.Count() > 0) ? (prop.GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : prop.Name;
                                 var value = reader[$"{propname}"];
                                 if (value.GetType() == typeof(string))
                                     value = (value as string).Trim();
+                                if (value.GetType() == typeof(DBNull))
+                                    value = null;
                                 prop.SetValue(item, value);
                             }
 
@@ -143,55 +145,58 @@ namespace DBF
         }
 
         /// <summary>
-        /// Pushes the given item of type <typeparamref name="T"/> to the sql database
+        /// Pushes the given item of type <typeparamref name="TModel"/> to the sql database
         /// </summary>
-        /// <typeparam name="T">The model corresponding to the database table</typeparam>
+        /// <typeparam name="TModel">The model corresponding to the database table</typeparam>
         /// <param name="item">The item to push to the sql database</param>
         /// <returns></returns>
-        public async Task Push<T>(T item)
+        public async override Task Push<TModel>(TModel item)
         {
             await Task.Run(() =>
             {
                 // Get all the properties to insert into the database
                 List<(string, object)> dbvalues = new List<(string, object)>();
                 // TODO: Ignore the ignore properties
-                foreach (var prop in typeof(T).GetProperties())
+                foreach (var prop in item.GetType().GetProperties())
                     dbvalues.Add(((prop.CustomAttributes.Count() > 0) ? (prop.GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : prop.Name, prop.GetValue(item)));
 
                 // TODO: add constraint checking here
                 // ------------------------------------------------------------
                 // If the options don't allow unsafe code and the where is null
                 // Throw an excpetion
-                //if (!item.HasPrimaryKeyValue(Schema) && !Options.AllowUnsafe) throw new Exception("The DBContext does not allow unsafe code and the action that would be ran is considered unsafe\r\nAction: trying to delete without setting the where with primary key as unique identifier");
-                //if (!item.RequiredValuesCorrect(Schema)) throw new Exception("Required values cannot be null");
+                if (!Schema.HasPrimaryKeyValue(item)) throw new Exception("The DBContext does not allow unsafe code and the action that would be ran is considered unsafe\r\nAction: trying to delete without setting the where with primary key as unique identifier");
+                if (!Schema.RequiredValuesCorrect(item)) throw new Exception("Required values cannot be null");
 
                 // Generate the insert query
-                string table = (typeof(T).CustomAttributes.Count() > 0) ? (typeof(T).GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : typeof(T).Name;
+                string table = (item.GetType().CustomAttributes.Count() > 0) ? (item.GetType().GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : item.GetType().Name;
                 string columns = "";
                 string values = "";
                 int count = 0;
                 foreach (var val in dbvalues)
                 {
                     columns += $"{val.Item1}";
-                    switch (val.Item2.GetType().ToString())
-                    {
-                        case "System.Guid":
-                        case "System.String":
-                            values += $"'{val.Item2}'";
-                            break;
-                        case "System.Boolean":
-                            var v = (bool)val.Item2 ? 1 : 0;
-                            values += $"{v}";
-                            break;
-                        case "System.DateTimeOffset":
-                            string dtFormat = "yyyy-MM-dd HH:mm:ss.fffffff zzz";
-                            DateTimeOffset dt = DateTimeOffset.Parse(val.Item2.ToString());
-                            values += $"'{dt.ToString(dtFormat)}'";
-                            break;
-                        default:
-                            values += $"{val.Item2}";
-                            break;
-                    }
+                    if (val.Item2 == null)
+                        values += "NULL";
+                    else
+                        switch (val.Item2.GetType().ToString())
+                        {
+                            case "System.Guid":
+                            case "System.String":
+                                values += $"'{val.Item2}'";
+                                break;
+                            case "System.Boolean":
+                                var v = (bool)val.Item2 ? 1 : 0;
+                                values += $"{v}";
+                                break;
+                            case "System.DateTimeOffset":
+                                string dtFormat = "yyyy-MM-dd HH:mm:ss.fffffff zzz";
+                                DateTimeOffset dt = DateTimeOffset.Parse(val.Item2.ToString());
+                                values += $"'{dt.ToString(dtFormat)}'";
+                                break;
+                            default:
+                                values += $"{val.Item2}";
+                                break;
+                        }
                     if (++count == dbvalues.Count) continue;
                     columns += ", ";
                     values += ", ";
@@ -212,20 +217,20 @@ namespace DBF
         }
 
         /// <summary>
-        /// Removes the given item from the sql database
+        /// Removes the given item from the sql database based on the primary key value
         /// </summary>
-        /// <typeparam name="T">The model of the item corresponding to the table from the sql database</typeparam>
+        /// <typeparam name="TModel">The model of the item corresponding to the table from the sql database</typeparam>
         /// <param name="predicate">An <see cref="Action"/> to set the where clause for the removal of the item/items from the sql database</param>
         /// <returns></returns>
-        public async Task Remove<T>(Action<T> predicate) where T : class, new()
+        public async override Task Remove<TModel>(Action<TModel> predicate)
         {
             await Task.Run(() =>
             {
                 // Create the item to remove
-                T item = new T();
+                TModel item = new TModel();
                 predicate(item);
 
-                //bool HasPrimaryKey = false;
+                bool HasPrimaryKey = false;
 
                 // Get the where state from the predicate if given
                 string where = null;
@@ -235,7 +240,8 @@ namespace DBF
 
                     var value = prop.GetValue(item);
 
-                    //if (!HasPrimaryKey) HasPrimaryKey = prop.IsPrimaryKey(Schema, typeof(T));
+                    // Check the primary key
+                    if (!HasPrimaryKey) HasPrimaryKey = Schema.IsPrimaryKey(prop, typeof(TModel));
 
                     // If there is already a where clause
                     // Add a "AND"
@@ -257,15 +263,20 @@ namespace DBF
                             propEquater = $"{value}";
                             break;
                     }
+
+                    // Update the where clause
                     where += $"{propname} = {propEquater}";
+
+                    // Break from the iterator
+                    break;
                 }
 
                 // If the options don't allow unsafe code and the where is null
                 // Throw an excpetion
-                //if (!HasPrimaryKey && !Options.AllowUnsafe) throw new Exception("The DBContext does not allow unsafe code and the action that would be ran is considered unsafe\r\nAction: trying to delete without setting the where with primary key as unique identifier");
+                if (!HasPrimaryKey) throw new Exception("The DBContext does not allow unsafe code and the action that would be ran is considered unsafe\r\nAction: trying to delete without setting the where with primary key as unique identifier");
 
                 // Get the name of the database table
-                string table = (typeof(T).CustomAttributes.Count() > 0) ? (typeof(T).GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : typeof(T).Name;
+                string table = (item.GetType().CustomAttributes.Count() > 0) ? (item.GetType().GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : item.GetType().Name;
 
                 // Create the query
                 string query = $"DELETE FROM {table} WHERE {where ?? "1 = 1"}";
@@ -286,29 +297,34 @@ namespace DBF
         /// Updates an item on the sql database based on the predicate
         /// the new value comes from the update <see cref="Action"/>
         /// </summary>
-        /// <typeparam name="T">The model corresponding to the database table to update</typeparam>
+        /// <remarks>
+        ///     TODO: Find a better way to update the item on the database to any availlable value.
+        /// </remarks>
+        /// <typeparam name="TModel">The model corresponding to the database table to update</typeparam>
         /// <param name="predicate">The <see cref="Action"/> defining the where clause for the update statement</param>
-        /// <param name="update">The <see cref="Action"/> setting the new values for the item</param>
         /// <returns></returns>
-        public async Task Update<T>(Action<T> predicate) where T : class, new()
+        public async override Task Update<TModel>(Action<TModel> predicate)
         {
             await Task.Run(() =>
             {
                 // Get the name of the database table
-                string table = (typeof(T).CustomAttributes.Count() > 0) ? (typeof(T).GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : typeof(T).Name;
+                string table = (typeof(TModel).CustomAttributes.Count() > 0) ? (typeof(TModel).GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : typeof(TModel).Name;
 
                 // Generate the new values
                 string newValues = null;
-                T item = new T();
+                TModel item = new TModel();
                 predicate(item);
 
-                //if (!item.HasPrimaryKeyValue(Schema)) throw new Exception("In order to update an item, the item must have a value for the primary key and the primary will be used in the where statement only.");
+                if (!Schema.HasPrimaryKeyValue(item)) throw new Exception("In order to update an item, the item must have a value for the primary key and the primary will be used in the where statement only.");
 
+                // Updated property value assingment
                 foreach (var prop in item.GetType().GetProperties())
                 {
-                    if (prop.GetValue(item) == null) continue; // || prop.IsPrimaryKey(Schema, typeof(T))
-
+                    // Get the value
                     var value = prop.GetValue(item);
+
+                    // Check if the value is null or the primary key
+                    if (value == null || Schema.IsPrimaryKey(prop, typeof(TModel))) continue;
 
                     // If there is already a where clause
                     // Add a "AND"
@@ -316,20 +332,30 @@ namespace DBF
 
                     var propname = (prop.CustomAttributes.Count() > 0) ? (prop.GetCustomAttributes(typeof(DBName), false).First() as DBName)?.Name : prop.Name;
                     var newValue = "";
-                    switch (value.GetType().ToString())
-                    {
-                        case "System.Guid":
-                        case "System.String":
-                            newValue = $"'{value}'";
-                            break;
-                        case "System.Boolean":
-                            var v = (bool)value ? 1 : 0;
-                            newValue = $"{v}";
-                            break;
-                        default:
-                            newValue = $"{value}";
-                            break;
-                    }
+
+                    // Check if the value is the default value for the type, if so 
+                    if (value == prop.PropertyType.GetDefault())
+                        newValue = "NULL";
+                    else
+                        switch (value.GetType().ToString())
+                        {
+                            case "System.Guid":
+                            case "System.String":
+                                newValue = $"'{value}'";
+                                break;
+                            case "System.Boolean":
+                                var v = (bool)value ? 1 : 0;
+                                newValue = $"{v}";
+                                break;
+                            case "System.DateTimeOffset":
+                                string dtFormat = "yyyy-MM-dd HH:mm:ss.fffffff zzz";
+                                DateTimeOffset dt = DateTimeOffset.Parse(value.ToString());
+                                newValue = $"'{dt.ToString(dtFormat)}'";
+                                break;
+                            default:
+                                newValue = $"{value}";
+                                break;
+                        }
                     newValues += $"{propname} = {newValue}";
                 }
 
@@ -337,7 +363,7 @@ namespace DBF
                 string where = null;
                 foreach (var prop in item.GetType().GetProperties())
                 {
-                    //if (!prop.IsPrimaryKey(Schema, typeof(T))) continue;
+                    if (!Schema.IsPrimaryKey(prop, typeof(TModel))) continue;
                     if (prop.GetValue(item) == null) continue;
 
                     var value = prop.GetValue(item);
@@ -366,7 +392,8 @@ namespace DBF
                 }
 
                 // Update statement must have a where clause
-                if (where == null) throw new Exception("Where clause in Update cannot be null");
+                if (string.IsNullOrWhiteSpace(where)) throw new Exception("Where clause in Update cannot be null");
+                if (string.IsNullOrWhiteSpace(newValues)) throw new Exception("Set clause in Update cannot be null");
 
                 // Generate the query
                 string query = $"UPDATE {table} SET {newValues} WHERE {where}";
